requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-groq>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2
openai>=1.3.7
numpy>=1.26.0,<2.0.0
pandas>=2.1.3
